cleanup compile and fixes.
fix out of bounds in read.

	modified:   build-arm-linux-gnueabihf/cracker
	new file:   cracker.log
	modified:   include/cracker.h
	modified:   include/cracker_arm_ir.h
	modified:   source/cracker_arm.c
	modified:   source/cracker_main.c

+++ include/cracker.h
@@ -113,7 +113,7 @@ typedef struct cracker_t {
 
 /* **** */
 
-uint32_t _read(cracker_p cj, uint32_t pat, uint8_t size);
+uint32_t _read(cracker_p cj, uint32_t pat, size_t size);
 symbol_p cracker_data(cracker_p cj, uint32_t pat, size_t size);
 void cracker_reg_dst(cracker_p cj, uint8_t r);
 void cracker_reg_src(cracker_p cj, uint8_t r);
